requirements.txt
Telethon==1.24.0
Pillow==8.4.0